+++ tests/Phan/PHP80Test.php
@@ -18,6 +18,7 @@ final class PHP80Test extends AbstractPhanFileTest
         'unused_variable_detection' => true,  // for use with tests of arrow functions
         'redundant_condition_detection' => true,  // for use with typed properties
         'target_php_version' => '8.0',
+        'minimum_target_php_version' => '7.4',  // test compatibility warnings for projects
         'plugins' => [
             'UseReturnValuePlugin',
             'UnreachableCodePlugin',